Added Cascading to Prisma
Added cascading deletes to the prisma schema to allow for the children tables to also be deleted when the main table dies. Then adjusted the delete button to allow for the deletion of multuple locations at once. I also modified some of the names in the tables to make more sense